amended app models

+++ transactions/models.py
@@ -4,7 +4,9 @@ from django.db import models
 # Create your models here.
 
 class Transactions(models.Model):
-    TransactionID=models.BigIntegerField(primary_key=True,null=False,blank=False)
+    TransactionID=models.AutoField(primary_key=True,null=False,blank=False)
+
+    #transaction row will only refer to one account: there should be two rows for trnasactions involving two clients of the bank.
     AccountID=models.BigIntegerField(null=False,blank=False)
 
     BankSortCode=models.CharField(max_length=8,null=False,blank=False) #requires validation: CHECK (SortCode LIKE '[0-9][0-9]-[0-9][0-9]-[0-9][0-9]')
